Staging

+++ app/Http/Controllers/AdminVideosController.php
@@ -332,8 +332,8 @@ if($row->active == 0){ $active = "Pending" ;$class="bg-warning"; }elseif($row->a
                     $Mobile_image =  'Mobile-default_image.jpg' ;
                     $Tablet_image =  'Tablet-default_image.jpg' ;
                                 
-                    Image::make($PC_image_path)->fit(720,1440)->save(base_path().'/public/uploads/images/'.$Mobile_image );
-                    Image::make($PC_image_path)->fit(360,960)->save(base_path().'/public/uploads/images/'.$Tablet_image );
+                    Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Mobile_image );
+                    Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Tablet_image );
                 
                     $video->mobile_image  = $Mobile_image;
                     $video->tablet_image  = $Tablet_image;
@@ -391,8 +391,8 @@ if($row->active == 0){ $active = "Pending" ;$class="bg-warning"; }elseif($row->a
                      $Mobile_image =  'Mobile-default_image.jpg' ;
                      $Tablet_image =  'Tablet-default_image.jpg' ;
                                  
-                     Image::make($PC_image_path)->fit(720,1440)->save(base_path().'/public/uploads/images/'.$Mobile_image );
-                     Image::make($PC_image_path)->fit(360,960)->save(base_path().'/public/uploads/images/'.$Tablet_image );
+                     Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Mobile_image );
+                     Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Tablet_image );
                  
                      $video->mobile_image  = $Mobile_image;
                      $video->tablet_image  = $Tablet_image;
@@ -452,8 +452,8 @@ if($row->active == 0){ $active = "Pending" ;$class="bg-warning"; }elseif($row->a
                     $Mobile_image =  'Mobile-default_image.jpg' ;
                     $Tablet_image =  'Tablet-default_image.jpg' ;
                                 
-                    Image::make($PC_image_path)->fit(720,1440)->save(base_path().'/public/uploads/images/'.$Mobile_image );
-                    Image::make($PC_image_path)->fit(360,960)->save(base_path().'/public/uploads/images/'.$Tablet_image );
+                    Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Mobile_image );
+                    Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Tablet_image );
                 
                     $video->mobile_image  = $Mobile_image;
                     $video->tablet_image  = $Tablet_image;
@@ -934,6 +934,7 @@ if(!empty($artistsdata)){
      */
     public function update(Request $request)
     {
+
          if (!Auth::user()->role == 'admin')
         {
             return redirect('/home');
@@ -995,25 +996,44 @@ if(!empty($artistsdata)){
 
         $path = public_path().'/uploads/videos/';
 
-           if(!empty($trailer)) {   
-            //code for remove old file
-            if($trailer != ''  && $trailer != null){
-                 $file_old = $path.$trailer;
-                if (file_exists($file_old)){
-                 unlink($file_old);
+
+        $video->trailer_type = $data['trailer_type'];
+
+        // dd($data);
+
+        if($data['trailer_type'] == 'video_mp4'){
+            if(!empty($trailer)) {   
+                //code for remove old file
+                if($trailer != ''  && $trailer != null){
+                     $file_old = $path.$trailer;
+                    if (file_exists($file_old)){
+                     unlink($file_old);
+                    }
                 }
-            }
-            //upload new file
-            $randval = Str::random(16);
-            $file = $trailer;
-            $trailer_vid  = $randval.'.'.$request->file('trailer')->extension();
-            $file->move($path, $trailer_vid);
-            $data['trailer']  = URL::to('/').'/public/uploads/videos/'.$trailer_vid;
-               $video->trailer = URL::to('/').'/public/uploads/videos/'.$trailer_vid;
-            } else {
-                $data['trailer'] = $video->trailer;
-            }  
+                //upload new file
+                $randval = Str::random(16);
+                $file = $trailer;
+                $trailer_vid  = $randval.'.'.$request->file('trailer')->extension();
+                $file->move($path, $trailer_vid);
+                $data['trailer']  = URL::to('/').'/public/uploads/videos/'.$trailer_vid;
+                   $video->trailer = URL::to('/').'/public/uploads/videos/'.$trailer_vid;
+                } else {
+                    $data['trailer'] = $video->trailer;
+                }  
+
+                }elseif($data['trailer_type'] == 'm3u8_url'){
+                    $video->trailer = $data['m3u8_trailer'];
+                }
+                elseif($data['trailer_type'] == 'mp4_url'){
+                    $video->trailer = $data['mp4_trailer'];
+                }
+                elseif($data['trailer_type'] == 'embed_url'){
+                    $video->trailer = $data['embed_trailer'];
+                }
         
+
+
+           
            $update_mp4 = $request->get('video');
            if(empty($data['active'])){
             $active = 0;
@@ -1151,9 +1171,9 @@ if(!empty($artistsdata)){
               $Mobile_image =  'Mobile'.$filename ;
               $Tablet_image =  'Tablet'.$filename ;
               
-              Image::make($file)->fit(720,1080)->save(base_path().'/public/uploads/images/'.$PC_image,80 );
-              Image::make($file)->fit(720,1440)->save(base_path().'/public/uploads/images/'.$Mobile_image,80 );
-              Image::make($file)->fit(360,960)->save(base_path().'/public/uploads/images/'.$Tablet_image,80 );
+              Image::make($file)->save(base_path().'/public/uploads/images/'.$PC_image,80 );
+              Image::make($file)->save(base_path().'/public/uploads/images/'.$Mobile_image,80 );
+              Image::make($file)->save(base_path().'/public/uploads/images/'.$Tablet_image,80 );
 
              $video->image  = $PC_image;
              $video->mobile_image  = $Mobile_image;
@@ -1762,9 +1782,9 @@ if(!empty($artistsdata)){
                   $Tablet_image =  'Tablet'.$filename ;
 
                   
-                  Image::make($files)->fit(720,1080)->save(base_path().'/public/uploads/images/'.$PC_image,80 );
-                  Image::make($files)->fit(720,1440)->save(base_path().'/public/uploads/images/'.$Mobile_image,80 );
-                  Image::make($files)->fit(360,960)->save(base_path().'/public/uploads/images/'.$Tablet_image,80 );
+                  Image::make($files)->save(base_path().'/public/uploads/images/'.$PC_image,80 );
+                  Image::make($files)->save(base_path().'/public/uploads/images/'.$Mobile_image,80 );
+                  Image::make($files)->save(base_path().'/public/uploads/images/'.$Tablet_image,80 );
 
                  $video->mobile_image  = $Mobile_image;
                  $video->tablet_image  = $Tablet_image;
@@ -1774,6 +1794,8 @@ if(!empty($artistsdata)){
                  $data['image'] = $video->image;
              }
             
+             $video->trailer_type = $data['trailer_type'];
+
             if($data['trailer_type'] == 'video_mp4'){
                 if($trailer != '') {   
                     //code for remove old file
@@ -2082,8 +2104,8 @@ if(!empty($artistsdata)){
                 $Mobile_image =  'Mobile-default_image.jpg' ;
                 $Tablet_image =  'Tablet-default_image.jpg' ;
                                             
-                Image::make($PC_image_path)->fit(720,1440)->save(base_path().'/public/uploads/images/'.$Mobile_image );
-                Image::make($PC_image_path)->fit(360,960)->save(base_path().'/public/uploads/images/'.$Tablet_image );
+                Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Mobile_image );
+                Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Tablet_image );
                             
                 $video->mobile_image  = $Mobile_image;
                 $video->tablet_image  = $Tablet_image;
@@ -2129,8 +2151,8 @@ if(!empty($artistsdata)){
                 $Mobile_image =  'Mobile-default_image.jpg' ;
                 $Tablet_image =  'Tablet-default_image.jpg' ;
                                             
-                Image::make($PC_image_path)->fit(720,1440)->save(base_path().'/public/uploads/images/'.$Mobile_image );
-                Image::make($PC_image_path)->fit(360,960)->save(base_path().'/public/uploads/images/'.$Tablet_image );
+                Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Mobile_image );
+                Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Tablet_image );
                             
                 $video->mobile_image  = $Mobile_image;
                 $video->tablet_image  = $Tablet_image;
@@ -2181,8 +2203,8 @@ if(!empty($artistsdata)){
                 $Mobile_image =  'Mobile-default_image.jpg' ;
                 $Tablet_image =  'Tablet-default_image.jpg' ;
                                             
-                Image::make($PC_image_path)->fit(720,1440)->save(base_path().'/public/uploads/images/'.$Mobile_image );
-                Image::make($PC_image_path)->fit(360,960)->save(base_path().'/public/uploads/images/'.$Tablet_image );
+                Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Mobile_image );
+                Image::make($PC_image_path)->save(base_path().'/public/uploads/images/'.$Tablet_image );
                             
                 $video->mobile_image  = $Mobile_image;
                 $video->tablet_image  = $Tablet_image;
